[DEV] Item 인터페이스 정의
