+++ controller/cart-controller.js
@@ -53,7 +53,7 @@ var getCartList = function (req, res, next) {
                 case 2:
                     result = _a.sent();
                     result.message = 'Success';
-                    logger_1.default.reportResponse(req.url, req.method, result);
+                    logger_1.default.reportResponse(req.url, req.method, result.message);
                     res.status(200).json(result);
                     return [3 /*break*/, 4];
                 case 3:
@@ -83,7 +83,7 @@ var addCart = function (req, res, next) {
                 case 2:
                     result = _b.sent();
                     result.message = "Success";
-                    logger_1.default.reportResponse(req.url, req.method, result);
+                    logger_1.default.reportResponse(req.url, req.method, result.message);
                     res.status(200).json(result);
                     return [3 /*break*/, 4];
                 case 3:
