--- data/dbAccess/interface.ts
@@ -1,14 +1,0 @@
-export interface Book {
-    id: number;
-    title: string;
-    name: string;
-    form: string;
-    isbn: string;
-    summary: string;
-    author: string;
-    pages: number;
-    contents: string;
-    pub_date: Date;
-    detail: string;
-    img: string;
-}
