+++ services/order-service.js
@@ -66,6 +66,7 @@ exports.default = {
                     case 0:
                         _a.trys.push([0, 2, , 3]);
                         bookIds = body.items.map(function (cur) { return (cur.bookId); });
+                        console.log(bookIds);
                         return [4 /*yield*/, dbAccess_1.orderDb.insertOrderAndDeleteCart(userId, body, bookIds)];
                     case 1:
                         _a.sent();
