+++ shared/type.ts
@@ -22,3 +22,23 @@ export interface UserInfo {
     address: string,
     contact: string,
 }
+
+export interface Book {
+    id: number;
+    title: string;
+    name: string;
+    form: string;
+    isbn: string;
+    summary: string;
+    author: string;
+    pages: number;
+    contents: string;
+    pub_date: Date;
+    detail: string;
+    img: string;
+}
+
+export interface Item {
+    bookId: number,
+    count: number
+}
